code is running, work on sample inputs to finish the project.

+++ Final Project-Routing with Graph Algorithms/Headers/Point.h
@@ -3,10 +3,10 @@
 class Point {
 public:
 	int id;
-	int x;
-	int y;
+	double x;
+	double y;
 
-	Point(int id, int x, int y) {
+	Point(int id, double x, double y) {
 		this->id = id;
 		this->x = x;
 		this->y = y;
